// generated/src/app/api-models/claim-hold-rules.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ClaimHoldRules {
  seqClhldRule: number;
  claimType: string;
  holdRule: string;
  reasonCode: string;
  description: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}